mv DGAdvection to framework and some cleanups #7603
