+++ framework/include/dgkernels/DGAdvection.h
@@ -1,0 +1,37 @@
+/****************************************************************/
+/*               DO NOT MODIFY THIS HEADER                      */
+/* MOOSE - Multiphysics Object Oriented Simulation Environment  */
+/*                                                              */
+/*           (c) 2010 Battelle Energy Alliance, LLC             */
+/*                   ALL RIGHTS RESERVED                        */
+/*                                                              */
+/*          Prepared by Battelle Energy Alliance, LLC           */
+/*            Under Contract No. DE-AC07-05ID14517              */
+/*            With the U. S. Department of Energy               */
+/*                                                              */
+/*            See COPYRIGHT for full restrictions               */
+/****************************************************************/
+
+#ifndef DGADVECTION_H
+#define DGADVECTION_H
+
+#include "DGKernel.h"
+
+class DGAdvection;
+
+template<>
+InputParameters validParams<DGAdvection>();
+
+class DGAdvection : public DGKernel
+{
+public:
+  DGAdvection(const InputParameters & parameters);
+
+protected:
+  virtual Real computeQpResidual(Moose::DGResidualType type);
+  virtual Real computeQpJacobian(Moose::DGJacobianType type);
+
+  RealVectorValue _velocity;
+};
+
+#endif //DGADVECTION_H
