+++ framework/src/dgkernels/DGAdvection.C
@@ -1,0 +1,90 @@
+/****************************************************************/
+/*               DO NOT MODIFY THIS HEADER                      */
+/* MOOSE - Multiphysics Object Oriented Simulation Environment  */
+/*                                                              */
+/*           (c) 2010 Battelle Energy Alliance, LLC             */
+/*                   ALL RIGHTS RESERVED                        */
+/*                                                              */
+/*          Prepared by Battelle Energy Alliance, LLC           */
+/*            Under Contract No. DE-AC07-05ID14517              */
+/*            With the U. S. Department of Energy               */
+/*                                                              */
+/*            See COPYRIGHT for full restrictions               */
+/****************************************************************/
+
+#include "DGAdvection.h"
+
+template<>
+InputParameters validParams<DGAdvection>()
+{
+  InputParameters params = validParams<DGKernel>();
+  params.addRequiredParam<RealVectorValue>("velocity", "Velocity vector");
+  params.addClassDescription("DG upwinding for the advection");
+  return params;
+}
+
+DGAdvection::DGAdvection(const InputParameters & parameters) :
+    DGKernel(parameters),
+    _velocity(getParam<RealVectorValue>("velocity"))
+{
+}
+
+Real
+DGAdvection::computeQpResidual(Moose::DGResidualType type)
+{
+  Real r = 0;
+
+  Real vdotn = _velocity * _normals[_qp];
+
+  switch (type)
+  {
+    case Moose::Element:
+      if (vdotn >= 0)
+        r += vdotn * _u[_qp] * _test[_i][_qp];
+      else
+        r += vdotn * _u_neighbor[_qp] * _test[_i][_qp];
+      break;
+
+    case Moose::Neighbor:
+      if (vdotn >= 0)
+        r -= vdotn * _u[_qp] * _test_neighbor[_i][_qp];
+      else
+        r -= vdotn * _u_neighbor[_qp] * _test_neighbor[_i][_qp];
+      break;
+  }
+
+  return r;
+}
+
+Real
+DGAdvection::computeQpJacobian(Moose::DGJacobianType type)
+{
+  Real r = 0;
+
+  Real vdotn = _velocity * _normals[_qp];
+
+  switch (type)
+  {
+    case Moose::ElementElement:
+      if (vdotn >= 0)
+        r += vdotn * _phi[_j][_qp] * _test[_i][_qp];
+      break;
+
+    case Moose::ElementNeighbor:
+      if (vdotn < 0)
+        r += vdotn * _phi_neighbor[_j][_qp] * _test[_i][_qp];
+      break;
+
+    case Moose::NeighborElement:
+      if (vdotn >= 0)
+        r -= vdotn * _phi[_j][_qp] * _test_neighbor[_i][_qp];
+      break;
+
+    case Moose::NeighborNeighbor:
+      if (vdotn < 0)
+        r -= vdotn * _phi_neighbor[_j][_qp] * _test_neighbor[_i][_qp];
+      break;
+  }
+
+  return r;
+}
