+++ test/src/base/MooseTestApp.C
@@ -153,7 +153,6 @@
 #include "ThrowMaterial.h"
 
 #include "DGMatDiffusion.h"
-#include "DGAdvection.h"
 #include "DGMDDBC.h"
 #include "DGFunctionConvectionDirichletBC.h"
 #include "CoupledKernelGradBC.h"
@@ -392,7 +391,6 @@ MooseTestApp::registerObjects(Factory & factory)
 
   // DG kernels
   registerDGKernel(DGMatDiffusion);
-  registerDGKernel(DGAdvection);
 
   // Interface kernels
   registerInterfaceKernel(InterfaceDiffusion);
@@ -453,7 +451,6 @@ MooseTestApp::registerObjects(Factory & factory)
   registerMaterial(RecomputeMaterial);
   registerMaterial(NewtonMaterial);
   registerMaterial(ThrowMaterial);
-
 
   registerScalarKernel(ExplicitODE);
   registerScalarKernel(ImplicitODEx);
